Crate iota_archival

Modules

reader

writer

Structs

CheckpointUpdates

FileMetadata

ManifestV1

Enums

FileType

Manifest

Constants

CHECKPOINT_FILE_MAGIC

Checkpoints and summaries are persisted as blob files. Files are committed to local store by duration or file size. Committed files are synced with the remote store continuously. Files are optionally compressed with the zstd compression format. Filenames follow the format <checkpoint_seq_num>. where checkpoint_seq_num is the first checkpoint present in that file. MANIFEST is the index and source of truth for all files present in the archive.

SUMMARY_FILE_MAGIC

Functions

create_file_metadata

create_file_metadata_from_bytes

finalize_manifest

Computes the SHA3 checksum of the Manifest and writes it to a byte vector.

read_manifest

Reads the manifest file from the store.

read_manifest_as_json

read_manifest_from_bytes

Reads the manifest file from the given byte vector and verifies the integrity of the file.

verify_archive_with_checksums

verify_archive_with_genesis_config

Loads and stores states from genesis then verifies and stores the checkpoints from the archive store.

verify_archive_with_local_store

Verifies the archive store by reading the checkpoints from the remote store and storing them in the local one.

write_manifest

Writes the Manifest to the remote store.

write_manifest_from_json